requirements.txt
pymupdf==1.23.10
dashscope==1.15.0
fastapi==0.104.0
uvicorn==0.29.0
python-dotenv==1.0.0